# StoredTableColumnAlignmentRule