--- a/src/lib/optimizer/strategy/stored_table_column_alignment_rule.cpp
+++ b/src/lib/optimizer/strategy/stored_table_column_alignment_rule.cpp
@@ -2,6 +2,9 @@
 
 #include <algorithm>
 #include <memory>
+#include <optional>
+#include <unordered_set>
+#include <vector>
 
 #include "logical_query_plan/abstract_lqp_node.hpp"
 #include "logical_query_plan/lqp_utils.hpp"
@@ -29,52 +32,56 @@ struct StoredTableNodeSharedPtrEqual final {
   }
 };
 
-template <typename Value>
-using ColumnPruningAgnosticStoredTableNodeMap =
-    std::unordered_map<std::shared_ptr<opossum::StoredTableNode>, Value, StoredTableNodeSharedPtrHash,
-                       StoredTableNodeSharedPtrEqual>;
+using ColumnPruningAgnosticSet = std::unordered_set<std::shared_ptr<opossum::StoredTableNode>,
+                                                    StoredTableNodeSharedPtrHash, StoredTableNodeSharedPtrEqual>;
+
+using ColumnPruningAgnosticMultiSet =
+    std::unordered_multiset<std::shared_ptr<opossum::StoredTableNode>, StoredTableNodeSharedPtrHash,
+                            StoredTableNodeSharedPtrEqual>;
 }  // namespace
 
 namespace opossum {
 
 void StoredTableColumnAlignmentRule::apply_to(const std::shared_ptr<AbstractLQPNode>& root) const {
-  ColumnPruningAgnosticStoredTableNodeMap<std::vector<std::shared_ptr<StoredTableNode>>>
-      gathered_stored_table_nodes_for_node;
-  ColumnPruningAgnosticStoredTableNodeMap<std::vector<ColumnID>> aligned_pruned_column_ids_for_node;
+  // stores exactly one representative for each StoredTableNode group. Members of the same StoredTableNode group
+  // have the same table name and pruned chunks.
+  auto group_representatives = ColumnPruningAgnosticSet{};
 
+  // stores all StoredTableNodes grouped by key.
+  auto grouped_stored_table_nodes = ColumnPruningAgnosticMultiSet{};
+
+  // iterate over the LQP and store all StoredTableNodes in multiple sets/groups: nodes of the same set/group have the
+  // same table name and the same pruned chunks.
   visit_lqp(root, [&](const auto& node) {
     if (node->type == LQPNodeType::StoredTable) {
       const auto stored_table_node = std::dynamic_pointer_cast<StoredTableNode>(node);
-
-      auto& gathered_stored_table_nodes = gathered_stored_table_nodes_for_node[stored_table_node];
-      gathered_stored_table_nodes.emplace_back(stored_table_node);
-
-      // update maximum superset of pruned column ids for a StoredTableNodes with the same table name and same set of
-      // pruned chunk ids
-      if (aligned_pruned_column_ids_for_node.contains(stored_table_node)) {
-        auto& aligned_pruned_column_ids = aligned_pruned_column_ids_for_node[stored_table_node];
-        std::vector<ColumnID> updated_pruned_column_ids{};
-        updated_pruned_column_ids.reserve(aligned_pruned_column_ids.size() +
-                                          stored_table_node->pruned_column_ids().size());
-        // sortedness of input containers is guaranteed by set_pruned_column_ids
-        std::set_intersection(aligned_pruned_column_ids.begin(), aligned_pruned_column_ids.end(),
-                              stored_table_node->pruned_column_ids().begin(),
-                              stored_table_node->pruned_column_ids().end(),
-                              std::back_inserter(updated_pruned_column_ids));
-        updated_pruned_column_ids.shrink_to_fit();
-        aligned_pruned_column_ids_for_node[stored_table_node] = updated_pruned_column_ids;
-      } else {
-        aligned_pruned_column_ids_for_node[stored_table_node] = stored_table_node->pruned_column_ids();
-      }
+      DebugAssert(stored_table_node, "LQPNode with type 'StoredTable' could not be casted to a StoredTableNode.");
+      group_representatives.emplace(stored_table_node);
+      grouped_stored_table_nodes.emplace(stored_table_node);
     }
     return LQPVisitation::VisitInputs;
   });
 
-  // iterate over groups of StoredTableNode and set the maximum superset of columns for each StoredTableNode in each
-  // StoredTableNode group. A group is characterized by the table name and the pruned chunk ids.
-  for (const auto& node_nodes_pair : gathered_stored_table_nodes_for_node) {
-    for (const auto& stored_table_node : node_nodes_pair.second) {
-      stored_table_node->set_pruned_column_ids(aligned_pruned_column_ids_for_node[stored_table_node]);
+  // for each group of StoredTableNodes, (1) iterate over the nodes and calculate the set intersection of pruned
+  // column ids and (2) iterate over the nodes and set the aligned pruned column ids.
+  for (const auto& group_representative : group_representatives) {
+    std::optional<std::vector<ColumnID>> aligned_pruned_column_ids;
+    auto group_range = grouped_stored_table_nodes.equal_range(group_representative);
+    for (auto iter = group_range.first; iter != group_range.second; ++iter) {
+      const auto& stored_table_node = *iter;
+      if (!aligned_pruned_column_ids) {
+        aligned_pruned_column_ids = stored_table_node->pruned_column_ids();
+      } else {
+        std::vector<ColumnID> updated_pruned_column_ids{};
+        std::set_intersection(aligned_pruned_column_ids->begin(), aligned_pruned_column_ids->end(),
+                              stored_table_node->pruned_column_ids().begin(),
+                              stored_table_node->pruned_column_ids().end(),
+                              std::back_inserter(updated_pruned_column_ids));
+        aligned_pruned_column_ids = std::move(updated_pruned_column_ids);
+      }
+    }
+    for (auto iter = group_range.first; iter != group_range.second; ++iter) {
+      (*iter)->set_pruned_column_ids(*aligned_pruned_column_ids);
     }
   }
 }