… src/lib/optimizer/strategy/stored_table_column_alignment_rule.hpp
… src/test/optimizer/strategy/stored_table_column_alignment_rule_test.cpp