## Description
The PQP sub plan memoization for StoredTableNodes is sensitive to the node's table name, set of pruned chunks and set of pruned columns. Consequently, for multiple nodes with the same table name, same pruned chunks and same pruned columns, only one GetTable operator is created and executed. In some queries, the ColumnPruningRule and ChunkPruningRule provide an LQP with multiple StoredTableNodes where the table names and sets of pruned chunks are equal but the sets of pruned columns are different. This leads to the creation and execution of different GetTable operators instead of only one operator.

This rule identifies StoredTableNodes that differ only in the pruned columns. It then intersects the lists of pruned columns. While this means that some columns are left unpruned, it makes the job easier for the memoization in the LQPTranslator. In our experiments, this has led to significant performance improvements and negligible reductions.

## Benchmark Comparison
comparison: d43fb623d vs f7d6246ec
executed on pella, NUMA node 1
build: RelWithDebInfo

### TPC-H
`./hyriseBenchmarkTPCH -s 1 -r 30`
```
+----------------+----------------+------+----------------+------+------------+----------------------+
| Benchmark      | prev. iter/s   | runs | new iter/s     | runs | change [%] |              p-value |
+----------------+----------------+------+----------------+------+------------+----------------------+
| TPC-H 01       | 0.808261215687 | 30   | 0.775613844395 | 30   | -4%        | (run time too short) |
| TPC-H 02       | 39.4202270508  | 30   | 40.5404472351  | 30   | +3%        | (run time too short) |
| TPC-H 03       | 5.11661243439  | 30   | 5.16426706314  | 30   | +1%        | (run time too short) |
| TPC-H 04       | 6.22300100327  | 30   | 6.11033773422  | 30   | -2%        | (run time too short) |
| TPC-H 05       | 1.98856902122  | 30   | 1.98523318768  | 30   | -0%        | (run time too short) |
| TPC-H 06       | 92.9693603516  | 30   | 93.6799240112  | 30   | +1%        | (run time too short) |
| TPC-H 07       | 7.18500375748  | 30   | 7.23907995224  | 30   | +1%        | (run time too short) |
| TPC-H 08       | 7.44914722443  | 30   | 7.46293115616  | 30   | +0%        | (run time too short) |
| TPC-H 09       | 1.05010020733  | 30   | 1.04464066029  | 30   | -1%        | (run time too short) |
| TPC-H 10       | 1.95173037052  | 30   | 1.94143795967  | 30   | -1%        | (run time too short) |
| TPC-H 11       | 51.842124939   | 30   | 51.6629066467  | 30   | -0%        | (run time too short) |
| TPC-H 12       | 10.4508676529  | 30   | 9.9458770752   | 30   | -5%        | (run time too short) |
| TPC-H 13       | 1.52068531513  | 30   | 1.51268649101  | 30   | -1%        | (run time too short) |
| TPC-H 14       | 25.84806633    | 30   | 25.5330581665  | 30   | -1%        | (run time too short) |
| TPC-H 15       | 56.921875      | 30   | 57.5614852905  | 30   | +1%        | (run time too short) |
| TPC-H 16       | 5.95111179352  | 30   | 6.00333356857  | 30   | +1%        | (run time too short) |
| TPC-H 17       | 9.75971412659  | 30   | 16.4921894073  | 30   | +69%       | (run time too short) |
| TPC-H 18       | 0.735659241676 | 30   | 0.73869240284  | 30   | +0%        | (run time too short) |
| TPC-H 19       | 12.1287403107  | 30   | 12.2900209427  | 30   | +1%        | (run time too short) |
| TPC-H 20       | 28.2435932159  | 30   | 28.4328689575  | 30   | +1%        | (run time too short) |
| TPC-H 21       | 0.821533858776 | 30   | 1.16572213173  | 30   | +42%       | (run time too short) |
| TPC-H 22       | 7.78854417801  | 30   | 7.86301755905  | 30   | +1%        | (run time too short) |
| geometric mean |                |      |                |      | +4%        |                      |
+----------------+----------------+------+----------------+------+------------+----------------------+
```

### TPC-DS
`./hyriseBenchmarkTPCDS -s 1 -r 30`
```
+----------------+---------------+------+---------------+------+------------+----------------------+
| Benchmark      | prev. iter/s  | runs | new iter/s    | runs | change [%] |              p-value |
+----------------+---------------+------+---------------+------+------------+----------------------+
| 01             | 14.8473272324 | 30   | 15.1353216171 | 30   | +2%        | (run time too short) |
| 03             | 57.7059631348 | 30   | 57.7753639221 | 30   | +0%        | (run time too short) |
| 06             | 20.8637065887 | 30   | 20.9599342346 | 30   | +0%        | (run time too short) |
| 07             | 6.79859972    | 30   | 6.51345396042 | 30   | -4%        | (run time too short) |
| 09             | 4.04866456985 | 30   | 4.03576564789 | 30   | -0%        | (run time too short) |
| 10             | 20.539100647  | 30   | 20.6436004639 | 30   | +1%        | (run time too short) |
| 13             | 5.34782171249 | 30   | 5.38502931595 | 30   | +1%        | (run time too short) |
| 15             | 41.746547699  | 30   | 41.9188842773 | 30   | +0%        | (run time too short) |
| 17             | 12.4100160599 | 30   | 12.5039396286 | 30   | +1%        | (run time too short) |
| 19             | 40.1315536499 | 30   | 40.2961769104 | 30   | +0%        | (run time too short) |
| 25             | 18.6570549011 | 30   | 18.9375076294 | 30   | +2%        | (run time too short) |
| 26             | 15.309504509  | 30   | 15.1728496552 | 30   | -1%        | (run time too short) |
| 28             | 7.99202489853 | 30   | 7.92670726776 | 30   | -1%        | (run time too short) |
| 29             | 12.2146492004 | 30   | 12.3524341583 | 30   | +1%        | (run time too short) |
| 31             | 3.84116411209 | 30   | 3.83518838882 | 30   | -0%        | (run time too short) |
| 34             | 17.777009964  | 30   | 17.8109416962 | 30   | +0%        | (run time too short) |
| 35             | 4.74193954468 | 30   | 4.78170204163 | 30   | +1%        | (run time too short) |
| 39a            | 2.4887316227  | 30   | 2.48633623123 | 30   | -0%        | (run time too short) |
| 39b            | 2.48768496513 | 30   | 2.49046421051 | 30   | +0%        | (run time too short) |
| 41             | 17.5549736023 | 30   | 16.9735965729 | 30   | -3%        | (run time too short) |
| 42             | 61.6618843079 | 30   | 61.7899055481 | 30   | +0%        | (run time too short) |
| 43             | 2.19654202461 | 30   | 2.19479179382 | 30   | -0%        | (run time too short) |
| 45             | 43.8250770569 | 30   | 43.9056663513 | 30   | +0%        | (run time too short) |
| 48             | 5.65509653091 | 30   | 5.6732635498  | 30   | +0%        | (run time too short) |
| 50             | 36.4451217651 | 30   | 36.749332428  | 30   | +1%        | (run time too short) |
| 52             | 60.7162437439 | 30   | 60.817199707  | 30   | +0%        | (run time too short) |
| 55             | 62.1214942932 | 30   | 62.0861740112 | 30   | -0%        | (run time too short) |
| 62             | 6.91570949554 | 30   | 6.84029006958 | 30   | -1%        | (run time too short) |
| 65             | 7.27852344513 | 30   | 7.18192815781 | 30   | -1%        | (run time too short) |
| 69             | 16.3616542816 | 30   | 16.3075428009 | 30   | -0%        | (run time too short) |
| 73             | 26.6281795502 | 30   | 26.4427642822 | 30   | -1%        | (run time too short) |
| 79             | 9.3133058548  | 30   | 9.25420951843 | 30   | -1%        | (run time too short) |
| 81             | 28.1001358032 | 30   | 28.116689682  | 30   | +0%        | (run time too short) |
| 83             | 51.9071273804 | 30   | 52.4082145691 | 30   | +1%        | (run time too short) |
| 85             | 6.56966876984 | 30   | 6.73189687729 | 30   | +2%        | (run time too short) |
| 88             | 4.78254747391 | 30   | 4.78843021393 | 30   | +0%        | (run time too short) |
| 91             | 30.44049263   | 30   | 30.6110363007 | 30   | +1%        | (run time too short) |
| 93             | 2.13753294945 | 30   | 2.11107897758 | 30   | -1%        | (run time too short) |
| 96             | 49.883934021  | 30   | 49.6257514954 | 30   | -1%        | (run time too short) |
| 97             | 1.30872070789 | 30   | 1.31454968452 | 30   | +0%        | (run time too short) |
| 99             | 3.34695076942 | 30   | 3.33235883713 | 30   | -0%        | (run time too short) |
| geometric mean |               |      |               |      | -0%        |                      |
+----------------+---------------+------+---------------+------+------------+----------------------+
```

### Testing
- [x] `./hyriseBenchmarkTPCH -s 0.01 -r 1 --verify` (f7d6246ec)
- [x] `./hyriseBenchmarkTPCDS -s 1 -r 1 --verify` (460d187)
- [x] `hyriseTest` (f7d6246ec)

### Query Plan Example: TPC-H Q17

Old LQP
![TPC-H_17-LQP_old](https://user-images.githubusercontent.com/20044656/82287557-bf359b80-99a0-11ea-8194-740d6c589722.png)

Old PQP
![TPC-H_17-PQP_old](https://user-images.githubusercontent.com/20044656/82287561-c3fa4f80-99a0-11ea-9208-1cb971e722c6.png)

New LQP
![TPC-H_17-LQP_new](https://user-images.githubusercontent.com/20044656/82287587-cb215d80-99a0-11ea-9257-782d0c09df7b.png)

New PQP
![TPC-H_17-PQP_new](https://user-images.githubusercontent.com/20044656/82287596-d1173e80-99a0-11ea-9d38-01d0bb9c0b84.png)