@mrks 2020-05-15
For the tests, I'd additionally check that the hash values and `operator==` are equal after the modifications. After all, that's what matters